# chore(latency-tests): reissue more notes