Reissuing is so fast now, that it's hard to bench any potential improvements. Issue more notes, to make it work harder.

In my tests, when using `ci` build profile:

Before:

LATENCY REISSUE: min: 0.6s, avg: 0.7s, median: 0.6s, p90: 0.7s, max: 0.7s, sum: 19.5s

Now:

LATENCY REISSUE: min: 1.3s, avg: 1.4s, median: 1.4s, p90: 1.5s, max: 1.5s, sum: 42.4s